requirements.txt
fastapi==0.115.7
motor==3.6.1
pydantic==2.10.5
python-dotenv==1.0.1
